pyproject.toml
[tool.poetry]
name = "movies-ugc-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pyjwt = "2.6.0"
cryptography = "40.0.2"
fastapi = "0.95.1"
clickhouse-driver = "0.2.6"
confluent-kafka = "2.1.1"
aiokafka = "0.8.0"
gunicorn = "20.1.0"
click = "8.1.3"
orjson = "3.8.12"
python-dotenv = "1.0.0"
dependency-injector = "4.41.0"
dpath = "2.1.5"


[tool.poetry.group.dev.dependencies]
# tests
pytest = "7.3.1"
pytest-asyncio = "^0.21.0"
# linters
black = "23.3.0"
isort = "5.12.0"
flake8 = "6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    .venv
    | venv
)/
'''

[tool.isort]
line_length = 79
combine_as_imports = true
include_trailing_comma = true
lines_after_imports = 2
skip = [".venv", "venv"]
